refactor + suspense
